优化记录代码结构
