+++ src/main/java/com/forteach/quiz/interaction/execute/dto/ExerciseBooksDto.java
@@ -1,6 +1,6 @@
 package com.forteach.quiz.interaction.execute.dto;
 
-import com.forteach.quiz.interaction.execute.domain.InteractQuestionsRecord;
+import com.forteach.quiz.interaction.execute.domain.record.InteractQuestionsRecord;
 
 import java.util.List;
