+++ src/main/java/com/forteach/quiz/interaction/execute/dto/QuestionsDto.java
@@ -1,6 +1,6 @@
 package com.forteach.quiz.interaction.execute.dto;
 
-import com.forteach.quiz.interaction.execute.domain.InteractQuestionsRecord;
+import com.forteach.quiz.interaction.execute.domain.record.InteractQuestionsRecord;
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
