+++ src/main/java/com/forteach/quiz/interaction/execute/service/ActiveInitiativeService.java
@@ -18,7 +18,6 @@ import org.springframework.data.redis.core.ReactiveStringRedisTemplate;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.List;
