+++ src/main/java/com/forteach/quiz/interaction/execute/service/ClassRoomService.java
@@ -2,7 +2,7 @@ package com.forteach.quiz.interaction.execute.service;
 
 import com.forteach.quiz.common.DefineCode;
 import com.forteach.quiz.common.MyAssert;
-import com.forteach.quiz.config.ExceptionHandlers;
+import com.forteach.quiz.interaction.execute.service.record.InteractRecordExecuteService;
 import com.forteach.quiz.service.StudentsService;
 import com.forteach.quiz.web.pojo.Students;
 import com.forteach.quiz.web.vo.InteractiveRoomVo;
