+++ src/main/java/com/forteach/quiz/interaction/execute/service/TaskInteractService.java
@@ -2,10 +2,11 @@ package com.forteach.quiz.interaction.execute.service;
 
 import com.forteach.quiz.exceptions.AskException;
 import com.forteach.quiz.interaction.execute.domain.ActivityAskAnswer;
+import com.forteach.quiz.interaction.execute.service.record.InsertInteractRecordService;
+import com.forteach.quiz.interaction.execute.service.record.InteractRecordExecuteService;
 import com.forteach.quiz.interaction.execute.web.vo.InteractiveSheetVo;
 import com.forteach.quiz.interaction.execute.web.vo.MoreGiveVo;
 import com.forteach.quiz.questionlibrary.domain.QuestionType;
-import com.forteach.quiz.questionlibrary.repository.TaskQuestionRepository;
 import com.mongodb.client.result.UpdateResult;
 import org.springframework.data.mongodb.core.ReactiveMongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
@@ -37,20 +38,20 @@ public class TaskInteractService {
 
     private final ReactiveMongoTemplate reactiveMongoTemplate;
 
-    private final TaskQuestionRepository questionRepository;
-
     private final InteractRecordExecuteService interactRecordExecuteService;
+
+    private final InsertInteractRecordService insertInteractRecordService;
 
     public TaskInteractService(ReactiveStringRedisTemplate stringRedisTemplate,
                                ReactiveHashOperations<String, String, String> reactiveHashOperations,
                                ReactiveMongoTemplate reactiveMongoTemplate,
-                               TaskQuestionRepository questionRepository,
+                               InsertInteractRecordService insertInteractRecordService,
                                InteractRecordExecuteService interactRecordExecuteService) {
         this.stringRedisTemplate = stringRedisTemplate;
         this.reactiveHashOperations = reactiveHashOperations;
         this.reactiveMongoTemplate = reactiveMongoTemplate;
-        this.questionRepository = questionRepository;
         this.interactRecordExecuteService = interactRecordExecuteService;
+        this.insertInteractRecordService = insertInteractRecordService;
     }
 
     /**
@@ -75,7 +76,7 @@ public class TaskInteractService {
 
         //TODO 未记录
         return Flux.concat(set, time, clearCut).filter(flag -> !flag)
-                .filterWhen(obj -> interactRecordExecuteService.releaseInteractRecord(giveVo.getCircleId(), giveVo.getQuestionId(), giveVo.getSelected(), giveVo.getCategory(), "interacts"))
+                .filterWhen(obj -> insertInteractRecordService.releaseInteractRecord(giveVo.getCircleId(), giveVo.getQuestionId(), giveVo.getSelected(), giveVo.getCategory(), "interacts"))
                 .count();
 
     }
@@ -115,7 +116,7 @@ public class TaskInteractService {
                 .filterWhen(set -> sendValue(sheetVo))
                 .filterWhen(right -> setRedis(sheetVo.getExamineeIsReplyKey(QuestionType.TaskQuestion), sheetVo.getExamineeId(), sheetVo.getAskKey(QuestionType.TaskQuestion)))
                 //保存到记录mongodb 集合
-                .filterWhen(vo -> interactRecordExecuteService.pushMongo(vo, "interacts"))
+                .filterWhen(vo -> insertInteractRecordService.pushMongo(vo, "interacts"))
                 .thenReturn(sheetVo.getCut());
     }
