+++ src/test/java/com/forteach/quiz/interaction/execute/service/InteractRecordExecuteServiceTest.java
@@ -1,12 +1,11 @@
 package com.forteach.quiz.interaction.execute.service;
 
+import com.forteach.quiz.interaction.execute.service.record.InteractRecordExecuteService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-
-import static org.junit.Assert.*;
 
 /**
  * @Auther: zhangyy
